refactor: data syncing key single fetch - optimized to the syncing mechanism to only check for the key once - change the function locations in their respective positions
