+++ src/lib/local/data/sync.ts
@@ -1,145 +1,120 @@
 import { Result } from "pg";
-import {
+import type {
   BuildingData,
-  ClassMapValue,
   CollegeData,
   DivisionData,
   DormData,
   RoomData,
+  TableSyncInfo,
 } from "../../types";
 import { getDB } from "./pgliteDB";
-import { getSyncKey, getSyncKeysFromLs, localTableSyncCheck, updateSyncKeyFromLs } from "./utils";
+import { getJSONFetch } from "./utils";
 import { syncToastStore } from "../../store.svelte";
 
-export async function getLocalBuildings(): Promise<BuildingData[] | undefined> {
+export function getSyncKeysFromLs(): {
+  [key: string]: string;
+} | null {
+  const lsStore = localStorage.getItem("sync-key");
+
+  if (lsStore === null) {
+    localStorage.setItem(
+      "sync-key",
+      `{
+      "buildings": "",
+      "colleges": "",
+      "divisions": "",
+      "rooms": "",
+      "dorms": "",
+      "classes": ""
+    }`,
+    );
+    return null;
+  }
+
   try {
-    const localDB = getDB();
-    await localDB.waitReady;
-    const data = (await localDB.query(`
-        SELECT building_name AS "buildingName", lon, lat, id, directions FROM buildings
-      `)) as Result<BuildingData>;
-    return data.rows;
+    const keys = JSON.parse(lsStore);
+    if (typeof keys === "object" && "buildings" in keys) {
+      return keys as {
+        [key: string]: string;
+      };
+    } else {
+      return null;
+    }
   } catch (e) {
-    console.error("Error: ", e);
-    return undefined;
+    console.error("Error: the sync keys in the localStorage was corrupted");
+    localStorage.setItem(
+      "sync-key",
+      `{
+      "buildings": "",
+      "colleges": "",
+      "divisions": "",
+      "rooms": "",
+      "dorms": "",
+      "classes": ""
+    }`,
+    );
+    return null;
   }
 }
 
-export async function getLocalColleges(): Promise<CollegeData[] | undefined> {
+export async function localTableSyncCheck(
+  tableName: string,
+): Promise<TableSyncInfo> {
   try {
-    const localDB = getDB();
-    await localDB.waitReady;
-    const data = (await localDB.query(`
-        SELECT college_name AS "collegeName", id FROM colleges;
-      `)) as Result<CollegeData>;
-    return data.rows;
+    const remoteSyncKey = await getSyncKey(tableName);
+    const syncKeyLs = getSyncKeysFromLs();
+    if (syncKeyLs === null)
+      return {
+        valid: false,
+        newKey: remoteSyncKey,
+      };
+    const tableSyncKey = syncKeyLs[tableName];
+    return {
+      valid:
+        typeof tableSyncKey === "string" &&
+        remoteSyncKey !== null &&
+        tableSyncKey === remoteSyncKey,
+      newKey: remoteSyncKey,
+    };
   } catch (e) {
-    console.error("Error: ", e);
-    return undefined;
+    console.error(e);
+    return {
+      valid: false,
+      newKey: null,
+    };
   }
 }
 
-export async function getLocalDivisions(): Promise<DivisionData[] | undefined> {
+export function updateSyncKeyFromLs(tableName: string, newKey: string) {
+  const syncKeys = getSyncKeysFromLs();
+  if (syncKeys === null) return;
+  syncKeys[tableName] = newKey;
+  localStorage.setItem("sync-key", JSON.stringify(syncKeys));
+}
+
+export async function getSyncKey(table: string) {
   try {
-    const localDB = getDB();
-    await localDB.waitReady;
-    const data = (await localDB.query(`
-        SELECT division_name AS "divisionName", id FROM divisions;
-      `)) as Result<DivisionData>;
-    return data.rows;
+    const tableOnline = await getJSONFetch<{
+      success: boolean;
+      error: string | null;
+      data: {
+        id: number;
+        tableName: string | null;
+        syncKey: string | null;
+      };
+    }>(`/api/check/${table}`);
+    return tableOnline.data.syncKey;
   } catch (e) {
-    console.error("Error: ", e);
-    return undefined;
+    console.error(e);
+    return null;
   }
 }
 
-export async function getLocalDorms(): Promise<DormData[] | undefined> {
-  try {
-    const localDB = getDB();
-    await localDB.waitReady;
-    const data = (await localDB.query(`
-      SELECT
-        id,
-        dorm_name AS "dormName",
-        short_name AS "shortName",
-        lat,
-        lon,
-        gender,
-        capacity,
-        managing_office AS "managingOffice",
-        contact_email AS "contactEmail",
-        amenities,
-        osm_link AS "osmLink",
-        description,
-        is_up_managed AS "isUpManaged",
-        price_range AS "priceRange",
-        contact_phone AS "contactPhone",
-        facebook_link AS "facebookLink"
-      FROM dorms;
-      `)) as Result<DormData>;
-    return data.rows;
-  } catch (e) {
-    console.error("Error: ", e);
-    return undefined;
-  }
-}
-
-// export async function getLocalRooms(): Promise<RoomData[] | undefined> {
-//   try {
-//     const localDB = getDB();
-
-//     await localDB.waitReady;
-//     const data = (await localDB.query(`
-//       SELECT
-//         r.room_code AS code,
-//         r.directions AS directions,
-//         json_build_object('name',b.building_name, 'lat', b.lat, 'lon', b.lon, 'directions', b.directions ) as building,
-//         c.college_name as "collegeName",
-//         d.division_name as "divisionName",
-//         r.building_id as "buildingId",
-//         r.college_id as "collegeId",
-//         r.division_id as "divisionId"
-//       FROM rooms AS r
-//       LEFT JOIN buildings AS b ON b.id = r.building_id
-//       LEFT JOIN colleges as c ON c.id = r.college_id
-//       LEFT JOIN divisions AS d ON d.id = r.division_id;
-//       `)) as Result<RoomData>;
-//     return data.rows;
-//   } catch (e) {
-//     console.error("Error: ", e);
-//     return undefined;
-//   }
-// }
-
-export async function getLocalClasses(): Promise<ClassMapValue[] | undefined> {
-  try {
-    const localDB = getDB();
-
-    await localDB.waitReady;
-    const data = (await localDB.query(`
-      SELECT
-        c.id,
-        c.course_code as "courseCode",
-        c.section,
-        c.type,
-        c.schedule,
-        c.directions,
-        c.course_title as "courseTitle",
-        c.term_id as "termId",
-        c.room_id as "roomId"
-      FROM classes AS c
-      LEFT JOIN rooms AS r ON r.id = c.room_id;
-    `)) as Result<ClassMapValue>;
-    return data.rows;
-  } catch (e) {
-    console.error("Error: ", e);
-    return undefined;
-  }
-}
-
-export async function syncBuildings(remoteBuildings: BuildingData[]) {
-  const res = await localTableSyncCheck("buildings");
-  if (res.valid) return;
+export async function syncBuildings(
+  checker: TableSyncInfo,
+  remoteBuildings: BuildingData[],
+) {
+  if (checker.valid) return;
   const localDB = getDB();
 
   await localDB.waitReady;
@@ -166,12 +141,14 @@ export async function syncBuildings(remoteBuildings: BuildingData[]) {
       console.error(e);
     }
   }
-  updateSyncKeyFromLs("buildings", res.newKey ?? "");
+  updateSyncKeyFromLs("buildings", checker.newKey ?? "");
 }
 
-export async function syncColleges(remoteColleges: CollegeData[]) {
-  const res = await localTableSyncCheck("colleges");
-  if (res.valid) return;
+export async function syncColleges(
+  checker: TableSyncInfo,
+  remoteColleges: CollegeData[],
+) {
+  if (checker.valid) return;
 
   const localDB = getDB();
 
@@ -197,12 +174,14 @@ export async function syncColleges(remoteColleges: CollegeData[]) {
     }
   }
 
-  updateSyncKeyFromLs("colleges", res.newKey ?? "");
+  updateSyncKeyFromLs("colleges", checker.newKey ?? "");
 }
 
-export async function syncDivisions(remoteDivisions: DivisionData[]) {
-  const res = await localTableSyncCheck("divisions");
-  if (res.valid) return;
+export async function syncDivisions(
+  checker: TableSyncInfo,
+  remoteDivisions: DivisionData[],
+) {
+  if (checker.valid) return;
 
   const localDB = getDB();
 
@@ -227,12 +206,14 @@ export async function syncDivisions(remoteDivisions: DivisionData[]) {
       console.error(e);
     }
   }
-  updateSyncKeyFromLs("divisions", res.newKey ?? "");
+  updateSyncKeyFromLs("divisions", checker.newKey ?? "");
 }
 
-export async function syncDorms(remoteDorms: DormData[]) {
-  const res = await localTableSyncCheck("dorms");
-  if (res.valid) return;
+export async function syncDorms(
+  checker: TableSyncInfo,
+  remoteDorms: DormData[],
+) {
+  if (checker.valid) return;
 
   const localDB = getDB();
 
@@ -286,7 +267,7 @@ export async function syncDorms(remoteDorms: DormData[]) {
       console.error(e);
     }
   }
-  updateSyncKeyFromLs("dorms", res.newKey ?? "");
+  updateSyncKeyFromLs("dorms", checker.newKey ?? "");
 }
 
 export async function resetBuildingsSyncStatus() {
@@ -300,19 +281,23 @@ export async function resetBuildingsSyncStatus() {
 }
 
 export async function localBuildingSyncStatus(id: number) {
-    try {
-        const localDB = getDB();
-        await localDB.waitReady;
-        const data = (await localDB.query(`
+  try {
+    const localDB = getDB();
+    await localDB.waitReady;
+    const data = (await localDB.query(
+      `
             SELECT rooms_fetched AS "roomsFetched"
             FROM buildings
             WHERE id = $1
-        `, [id])) as Result<{roomsFetched: boolean}>
-        if (data.rows.length === 0) throw new Error("Can't fetch a missing building row");
-        return data.rows[0];
-    }catch (e) {
-        console.error(e)
-    }
+        `,
+      [id],
+    )) as Result<{ roomsFetched: boolean }>;
+    if (data.rows.length === 0)
+      throw new Error("Can't fetch a missing building row");
+    return data.rows[0];
+  } catch (e) {
+    console.error(e);
+  }
 }
 
 export async function getLocalBuildingRooms(id: number) {
@@ -345,7 +330,7 @@ export async function getLocalBuildingRooms(id: number) {
   }
 }
 
-export async function checkLocalBuildingRoom (id: number) {
+export async function checkLocalBuildingRoom(id: number) {
   try {
     const remoteSyncKey = await getSyncKey("rooms");
     const syncKeyLs = getSyncKeysFromLs();
@@ -377,14 +362,14 @@ export async function checkLocalBuildingRoom (id: number) {
 }
 
 export async function syncBuildingRooms(id: number, rooms: RoomData[]) {
-    const synced = await checkLocalBuildingRoom(id);
+  const synced = await checkLocalBuildingRoom(id);
 
-    if (synced) return;
-    const localDB = getDB();
-    for (const room of rooms) {
-        try {
-          await localDB.query(
-            `
+  if (synced) return;
+  const localDB = getDB();
+  for (const room of rooms) {
+    try {
+      await localDB.query(
+        `
             INSERT INTO rooms (id, room_code, directions, building_id, college_id, division_id)
             VALUES ($1, $2, $3, $4, $5, $6)
             ON CONFLICT (id) DO UPDATE SET
@@ -395,13 +380,23 @@ export async function syncBuildingRooms(id: number, rooms: RoomData[]) {
             college_id = EXCLUDED.college_id,
             division_id = EXCLUDED.division_id;
             `,
-            [room.id, room.code, room.directions, room.buildingId, room.collegeId, room.divisionId],
-          );
-        } catch (e) {
-          console.error(e);
-        }
+        [
+          room.id,
+          room.code,
+          room.directions,
+          room.buildingId,
+          room.collegeId,
+          room.divisionId,
+        ],
+      );
+    } catch (e) {
+      console.error(e);
     }
-    await localDB.query(`UPDATE buildings SET rooms_fetched = true WHERE id = $1`, [id]);
+  }
+  await localDB.query(
+    `UPDATE buildings SET rooms_fetched = true WHERE id = $1`,
+    [id],
+  );
 }
 
 export async function resetCollegesSyncStatus() {
@@ -415,19 +410,23 @@ export async function resetCollegesSyncStatus() {
 }
 
 export async function localCollegeSyncStatus(id: number) {
-    try {
-        const localDB = getDB();
-        await localDB.waitReady;
-        const data = (await localDB.query(`
+  try {
+    const localDB = getDB();
+    await localDB.waitReady;
+    const data = (await localDB.query(
+      `
             SELECT rooms_fetched as "roomsFetched"
             FROM colleges
             WHERE id = $1;
-        `, [id])) as Result<{roomsFetched: boolean}>
-        if (data.rows.length === 0) throw new Error("Can't fetch a missing college row");
-        return data.rows[0];
-    } catch (e) {
-        console.error(e);
-    }
+        `,
+      [id],
+    )) as Result<{ roomsFetched: boolean }>;
+    if (data.rows.length === 0)
+      throw new Error("Can't fetch a missing college row");
+    return data.rows[0];
+  } catch (e) {
+    console.error(e);
+  }
 }
 
 export async function getLocalCollegeRooms(id: number) {
@@ -460,7 +459,7 @@ export async function getLocalCollegeRooms(id: number) {
   }
 }
 
-export async function checkLocalCollegeRoom (id: number) {
+export async function checkLocalCollegeRoom(id: number) {
   try {
     const remoteSyncKey = await getSyncKey("rooms");
     const syncKeyLs = getSyncKeysFromLs();
@@ -492,14 +491,14 @@ export async function checkLocalCollegeRoom (id: number) {
 }
 
 export async function syncCollegeRooms(id: number, rooms: RoomData[]) {
-    const synced = await checkLocalBuildingRoom(id);
+  const synced = await checkLocalBuildingRoom(id);
 
-    if (synced) return;
-    const localDB = getDB();
-    for (const room of rooms) {
-        try {
-          await localDB.query(
-            `
+  if (synced) return;
+  const localDB = getDB();
+  for (const room of rooms) {
+    try {
+      await localDB.query(
+        `
             INSERT INTO rooms (id, room_code, directions, building_id, college_id, division_id)
             VALUES ($1, $2, $3, $4, $5, $6)
             ON CONFLICT (id) DO UPDATE SET
@@ -510,13 +509,23 @@ export async function syncCollegeRooms(id: number, rooms: RoomData[]) {
             college_id = EXCLUDED.college_id,
             division_id = EXCLUDED.division_id;
             `,
-            [room.id, room.code, room.directions, room.buildingId, room.collegeId, room.divisionId],
-          );
-        } catch (e) {
-          console.error(e);
-        }
+        [
+          room.id,
+          room.code,
+          room.directions,
+          room.buildingId,
+          room.collegeId,
+          room.divisionId,
+        ],
+      );
+    } catch (e) {
+      console.error(e);
     }
-    await localDB.query(`UPDATE colleges SET rooms_fetched = true WHERE id = $1`, [id]);
+  }
+  await localDB.query(
+    `UPDATE colleges SET rooms_fetched = true WHERE id = $1`,
+    [id],
+  );
 }
 
 export async function resetDivisionsSyncStatus() {
@@ -530,19 +539,23 @@ export async function resetDivisionsSyncStatus() {
 }
 
 export async function localDivisionSyncStatus(id: number) {
-    try {
-        const localDB = getDB();
-        await localDB.waitReady;
-        const data = (await localDB.query(`
+  try {
+    const localDB = getDB();
+    await localDB.waitReady;
+    const data = (await localDB.query(
+      `
             SELECT rooms_fetched as "roomsFetched"
             FROM divisions
             WHERE id = $1;
-        `, [id])) as Result<{roomsFetched: boolean}>
-        if (data.rows.length === 0) throw new Error("Can't fetch a missing division row");
-        return data.rows[0];
-    } catch (e) {
-        console.error(e);
-    }
+        `,
+      [id],
+    )) as Result<{ roomsFetched: boolean }>;
+    if (data.rows.length === 0)
+      throw new Error("Can't fetch a missing division row");
+    return data.rows[0];
+  } catch (e) {
+    console.error(e);
+  }
 }
 
 export async function getLocalDivisionRooms(id: number) {
@@ -575,7 +588,7 @@ export async function getLocalDivisionRooms(id: number) {
   }
 }
 
-export async function checkLocalDivisionRoom (id: number) {
+export async function checkLocalDivisionRoom(id: number) {
   try {
     const remoteSyncKey = await getSyncKey("rooms");
     const syncKeyLs = getSyncKeysFromLs();
@@ -607,14 +620,14 @@ export async function checkLocalDivisionRoom (id: number) {
 }
 
 export async function syncDivisionRooms(id: number, rooms: RoomData[]) {
-    const synced = await checkLocalBuildingRoom(id);
+  const synced = await checkLocalBuildingRoom(id);
 
-    if (synced) return;
-    const localDB = getDB();
-    for (const room of rooms) {
-        try {
-          await localDB.query(
-            `
+  if (synced) return;
+  const localDB = getDB();
+  for (const room of rooms) {
+    try {
+      await localDB.query(
+        `
             INSERT INTO rooms (id, room_code, directions, building_id, college_id, division_id)
             VALUES ($1, $2, $3, $4, $5, $6)
             ON CONFLICT (id) DO UPDATE SET
@@ -625,11 +638,21 @@ export async function syncDivisionRooms(id: number, rooms: RoomData[]) {
             college_id = EXCLUDED.college_id,
             division_id = EXCLUDED.division_id;
             `,
-            [room.id, room.code, room.directions, room.buildingId, room.collegeId, room.divisionId],
-          );
-        } catch (e) {
-          console.error(e);
-        }
+        [
+          room.id,
+          room.code,
+          room.directions,
+          room.buildingId,
+          room.collegeId,
+          room.divisionId,
+        ],
+      );
+    } catch (e) {
+      console.error(e);
     }
-    await localDB.query(`UPDATE divisions SET rooms_fetched = true WHERE id = $1`, [id]);
+  }
+  await localDB.query(
+    `UPDATE divisions SET rooms_fetched = true WHERE id = $1`,
+    [id],
+  );
 }
